CC-5450 : Refactor Media Management (Classes/DB) in Airtime
working on generating playlists.
