+++ airtime_mvc/application/controllers/PlaylistController.php
@@ -2,6 +2,7 @@
 
 use Airtime\MediaItem\PlaylistQuery;
 use Airtime\MediaItem\Playlist;
+use Airtime\MediaItem\PlaylistPeer;
 
 class PlaylistController extends Zend_Controller_Action
 {
@@ -90,24 +91,45 @@ class PlaylistController extends Zend_Controller_Action
     
     public function clearAction()
     {
+    	$con = Propel::getConnection(PlaylistPeer::DATABASE_NAME);
+    	$con->beginTransaction();
+    	
     	try {
     		$playlist = $this->getPlaylist(); 
-    		$playlist->clear();
+    		$playlist->clearContent($con);
     		$this->createUpdateResponse($playlist);
+    		
+    		$con->commit();
     	}
     	catch (Exception $e) {
+    		$con->rollBack();
     		$this->view->error = $e->getMessage();
     	}
     }
     
     public function generateAction()
     {
+    	Logging::enablePropelLogging();
+    	
+    	$con = Propel::getConnection(PlaylistPeer::DATABASE_NAME);
+    	$con->beginTransaction();
+    	
     	try {
+
     		$playlist = $this->getPlaylist();
-    		$playlist->generate();
+    		$mediaIds = $playlist->generateContent($con);
+    		$playlist->addMedia($con, $mediaIds);
+    		$con->commit();
+    		
     		$this->createUpdateResponse($playlist);
+    		
+    		//$playlist->save($con);
+
+    		Logging::disablePropelLogging();
     	}
     	catch (Exception $e) {
+    		$con->rollBack();
+    		Logging::error($e->getFile().$e->getLine());
     		Logging::error($e->getMessage());
     		$this->view->error = $e->getMessage();
     	}
@@ -115,12 +137,18 @@ class PlaylistController extends Zend_Controller_Action
     
     public function shuffleAction()
     {
+    	$con = Propel::getConnection(PlaylistPeer::DATABASE_NAME);
+    	$con->beginTransaction();
+    	
     	try {
     		$playlist = $this->getPlaylist(); 
-    		$playlist->shuffle();
+    		$playlist->shuffleContent($con);
     		$this->createUpdateResponse($playlist);
+    		
+    		$con->commit();
     	}
     	catch (Exception $e) {
+    		$con->rollBack();
     		$this->view->error = $e->getMessage();
     	}
     }
