+++ airtime_mvc/application/models/airtime/Playlist.php
@@ -152,63 +152,4 @@ abstract class Playlist extends BasePlaylist implements \Interface_Playlistable
     
     	return $this;
     } // setRules()
-    
-    
-    public function savePlaylistContent($content, $replace=false)
-    {
-    	$con = Propel::getConnection(PlaylistPeer::DATABASE_NAME);
-		$con->beginTransaction();
-		
-		try {
-			
-			$m = array();
-			$currentContent = $this->getMediaContents(null, $con);
-			
-			if ($replace) {
-				$currentContent->delete($con);
-				$position = 0;
-			}
-			else {
-				$position = count($currentContent);
-			}
-			
-			foreach ($content as $item) {
-				
-				$mediaId = $item["id"];
-				$cuein = isset($item["cuein"]) ? $item["cuein"] : null;
-				$cueout = isset($item["cueout"]) ? $item["cueout"] : null;
-				$fadein = isset($item["fadein"]) ? $item["fadein"] : null;
-				$fadeout = isset($item["fadeout"]) ? $item["fadeout"] : null;
-				
-				$mediaItem = MediaItemQuery::create()->findPK($mediaId, $con);
-				$mediaContent = $this->buildContentItem($mediaItem, $position, $cuein, $cueout, $fadein, $fadeout);
-				$mediaContent->setPlaylist($this);
-				
-				$res = $mediaContent->validate();
-				if ($res === true) {
-					$m[] = $mediaContent;
-				}
-				else {
-					Logging::info($res);
-					throw new Exception("invalid media content");
-				}
-				
-				$position++;
-				
-				//save each content item in the transaction
-				//first so that Playlist preSave can calculate
-				//the new playlist length properly.
-				$mediaContent->save($con);
-			}
-			
-			$this->save($con);
-			
-			$con->commit();
-		}
-		catch (Exception $e) {
-			$con->rollBack();
-			Logging::error($e->getMessage());
-			throw $e;
-		}
-    }
 }
