+++ airtime_mvc/application/models/interfaces/Playlistable.php
@@ -2,9 +2,9 @@
 
 interface Interface_Playlistable
 {
-	public function generate();
-	public function shuffle();
-	public function clear();
+	public function generateContent(PropelPDO $con);
+	public function shuffleContent(PropelPDO $con);
+	public function clearContent(PropelPDO $con);
 	public function getContents();
 	public function getScheduledContent();
 }
